feat: enable adding an existing local kv db instead of always creating a new one
- Add `replaceExisting` option to `CreateKvStoreInput` that makes `kvStoresService.create`
function only replace an existing local kv database (`kv.sqlite3`) in the picked directory
when it set to 'true', otherwise the function will always reuse the existing local kv database
or creates a new one if there is no an already‑present one.

- Add a checkbox to `KvStoreForm.svelte` to give users the option to set the
`replaceExisting` option.

- In Kv Store Form, Hide `accessToken` field when the selected kv store type
isn't "remote", and hide `authToken` field when the selected type isn't "bridge".

+++ packages/preload/src/services/kvStoresService.ts
@@ -10,7 +10,9 @@ import { clearSavedParamsQuery } from '../db/browsingParamsQueries.js';
 export async function create(input: CreateKvStoreInput): Promise<boolean> {
     if (input.type == "local") {
         input.url = path.join(input.url, "kv.sqlite3")
-        await writeFile(input.url, "");
+        if (input.replaceExisting || !existsSync(input.url)) {
+            await writeFile(input.url, "");
+        }
     }
 
     const result = insertQuery.run(
